# Fix bug with an escaped slash