This PR fixes incorrect checking of Disallow: /*%2F rule