@dlecocq 2017-08-07
Thanks for your interest! C++ REP parsing is kind of a niche thing, and it's always fun, surprising when someone finds this little corner of the world.

I'm not sure I understand this use case -- is the goal to treat `%2F` specially? Other escaped entities are treated interchangeably with their unescaped versions (in `test-agent.cpp`, we have `EscapedRule`, `UnescapedRule`, `EscapedRuleWildcard`, `UnescapedRuleWildcard` to explicitly test that). It's entirely possible that provision is explicitly made for special treatment of `/`, but off hand I'm unaware of that.

My other concern is the use of `regex`. Before this PR, running `./bench`, gave me the approximately 3.8x the performance as with this PR on parsing the RFC-provided REP. That's currently the only benchmark that goes through the `Agent::escape` method:

```
# Without regex:
Benchmarking parse RFC with 100000 per run:
    Run 0: 801.7 ms
    Run 1: 793.409 ms
    Run 2: 811.514 ms
    Run 3: 802.514 ms
    Run 4: 782.877 ms
  Average: 798.403 ms
     Rate: 125.25 k-iter / s

# With regex:
Benchmarking parse RFC with 100000 per run:
    Run 0: 2876.68 ms
    Run 1: 3068.78 ms
    Run 2: 3132.61 ms
    Run 3: 3013.67 ms
    Run 4: 3093.32 ms
  Average: 3037.01 ms
     Rate: 32.9271 k-iter / s
```