@adeesent 2017-08-08
Thank you for your response! Yes, perfomance decreases dramatically and should find better solution. And thank you pointing out on ./bench, I totally missed it. 

This case apears to be important:
> MK1996 says, 'If a %xx encoded octet is encountered it is unencoded
> prior to comparison, unless it is the "/" character, which has
> special meaning in a path.'

http://www.robotstxt.org/norobots-rfc.txt