@adeesent 2017-08-08
E.g. we have rule
Disallow: /*%2F

So, in current implementation http://example.com/about is allowed, but http://example.com/about/ is not.