@dlecocq 2017-08-08
Interesting! I suppose I missed that section of the RFC initially.

The code internal to `Url` that deals with the escaping is here: https://github.com/seomoz/url-cpp/blob/master/src/url.cpp#L658 .

Originally, I thought it would be possible that code could be adapted or the interface of `Url#escape` could be extended to support providing a custom set of safe characters for each portion of the URL, enabling us to still use `Url#escape`, but explicitly provide configuration earmarking `/` as unsafe for use. I'm not sure whether that would work, but if it did it would probably involve using `strict` mode.

Still, it's something to investigate, and that `escape` code is relatively performant and is part of a code path that is currently used anyway.