@b4hand 2017-08-15
This is a very interesting edge case. I'm not sure exactly the best way to handle this is, but the recommendation by @dlecocq seems like a reasonable approach to investigate as a first pass. Alternatively, you could try doing the substitution manually using direct character replacements rather than using the regex replace which is probably doing multiple string allocations.