@b4hand 2017-09-06
After thinking about this issue a little more the 1996 RFC for REP does not actually support wildcards in the path. Wildcard support in REP was added later by Yahoo, Google, and other search engines, and I find no mention of special treatment of the escaped slash character in [Google's spec](https://developers.google.com/search/reference/robots_txt), so this particular edge case is odd in that it falls in between any of the public standards for REP.